OpenVLA package restructure v1
