+++ alliander_robotics/alliander_openvla/src/alliander_openvla/src_py/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-FileCopyrightText: Alliander N. V.
+#
+# SPDX-License-Identifier: Apache-2.0
